refactor(DST-1312): colocate demo files and remove standalone pattern routes
- Move filter components into examples/filter/ (from pattern/filter/)
- Move form components into examples/form/ (from pattern/form/)
- Move inventory components into examples/inventory/ (from inventory/)
- Remove standalone app-shell route (now served via /examples)
- Remove empty pattern/ and inventory/ directories
- Update "View Code" link in app-shell docs
