+++ docs/app/(examples)/examples/filter/applied-filter.tsx
