+++ docs/app/(examples)/examples/form/eventDetails.tsx
