+++ docs/app/(examples)/examples/form/registrationSettings.tsx
