+++ docs/app/(examples)/examples/inventory/Blog.tsx
