+++ docs/app/(examples)/examples/inventory/FAQ.tsx
