+++ docs/app/(examples)/examples/inventory/Inventory.tsx
