+++ docs/app/(examples)/examples/inventory/Newsletter.tsx
