+++ docs/app/(examples)/examples/inventory/OurMission.tsx
