+++ docs/app/(examples)/examples/inventory/Settings.tsx
