+++ docs/app/(examples)/examples/inventory/UserTable.tsx
