--- docs/app/(examples)/pattern/app-shell/(shell)/layout.tsx
@@ -1,11 +1,0 @@
-'use client';
-
-import type { PropsWithChildren } from 'react';
-import { ShellLayout } from '../../../_shared';
-import { config } from './navigation';
-
-const Layout = ({ children }: PropsWithChildren) => (
-  <ShellLayout config={config}>{children}</ShellLayout>
-);
-
-export default Layout;
