--- docs/app/(examples)/pattern/app-shell/(shell)/page.tsx
@@ -1,7 +1,0 @@
-'use client';
-
-import { DashboardContent } from '../../../../_shared/dashboard-content';
-
-const DashboardPage = () => <DashboardContent />;
-
-export default DashboardPage;
