+++ docs/content/patterns/app-shell/index.mdx
@@ -105,7 +105,7 @@ A complete App Shell with sidebar navigation (including drill-down), breadcrumbs
   <Center>[View Demo](/examples)</Center>
   <Center>
     [View
-    Code](https://github.com/marigold-ui/marigold/tree/main/docs/app/(examples)/pattern/app-shell)
+    Code](https://github.com/marigold-ui/marigold/tree/main/docs/app/(examples)/examples)
   </Center>
 </Stack>
